+++ utils.h
@@ -79,8 +79,8 @@ extern Objid get_system_object(const char *);
 
 extern int value_bytes(Var);
 
-extern const char *raw_bytes_to_moobinary(const char *buffer, int buflen);
-extern const char *moobinary_to_raw_bytes(const char *binary, int *rawlen);
+extern const char *raw_bytes_to_moobinary(const char *buffer, size_t buflen);
+extern const char *moobinary_to_raw_bytes(const char *binary, size_t *rawlen);
 
 #endif		/* !Utils_H */
